feat(gene): switches artworks grid on gene to be server-side rendered
